Safari-611.1.10/LayoutTests/http/tests/resourceLoadStatistics/sandboxed-iframe-redirect-ip-to-localhost-to-ip-expected.txt

Tests that redirects for sandboxed iframes get counted properly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testRunner.isStatisticsRegisteredAsSubFrameUnder("http://localhost", "http://127.0.0.1") is true
PASS testRunner.isStatisticsRegisteredAsRedirectingTo("http://localhost", "http://127.0.0.1") is true
PASS testRunner.isStatisticsRegisteredAsRedirectingTo("http://127.0.0.1", "http://localhost") is true
PASS successfullyParsed is true

TEST COMPLETE
